archive/containers/testing/requirements.txt
# Copyright (C) 2018-2023 Mark McIntyre

#cartopy INSTALL THIS WITH CONDA
# skip this
basemap
boto3
Cython
gmplot
google-auth
google-auth-oauthlib
google-api-python-client
googleapis-common-protos
jplephem
Keras
matplotlib
numpy
oauthlib
pandas
Pillow
pycparser
pyephem
pyproj 
PyQt5
pyswarms
pytz
PyYAML
scipy
Shapely
xmltodict
tweepy
fastparquet
cryptography
python-crontab
simplekml
astropy
gitpython
pymysql
mysql-connector-python
pre-commit
s3fs
# pyarrow
pytest
behave
pdoc
imageio
opencv-python